exporter/requirements.txt
# Core dependencies
prometheus-client==0.19.0
requests==2.31.0
pyyaml==6.0.1

# Development dependencies
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
black==23.12.0
flake8==6.1.0
mypy==1.7.1

# Type stubs
types-requests==2.31.0
types-PyYAML==6.0.12
